refactor: 스타일 폴더 적용
