+++ client/src/pages/Main.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { useSelector } from 'react-redux';
 import Login from './Login';
 import Home from './Home';
