+++ client/src/pages/Signup.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 
 import authApi from '../api/authApi';
-import styles from '../styles/Signup.module.css';
+import styles from '../styles/pages/Signup.module.css';
 import errorMessages from '../config/errorMessages';
 
 import Modal from '../components/Modal';
